+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   * dropped support for SVN repositories, only Git repositories are supported by now
   * implemented #10 replace herrera-io/version as it is not maintained any more
   * fixed #7 Git repository does not respect current branch
+  * added print of usage with `-h`, and `--help` command line options
+  * added print of current version with `-v` and `--version` command line options
 
 
 1.0.0 (2013-11-05)